+++ Magic/src/main/java/com/elmakers/mine/bukkit/protection/WorldGuardAPI.java
@@ -330,6 +330,19 @@ public class WorldGuardAPI {
         return null;
     }
 
+    @Nullable
+    public String getPortalWarp(Player player, Location location) {
+        if (location != null && worldGuardPlugin != null && customFlags != null) {
+            ApplicableRegionSet checkSet = getRegionSet(location);
+            if (checkSet == null) {
+                return null;
+            }
+
+            return customFlags.getPortalWarp(getAssociable(player), checkSet);
+        }
+        return null;
+    }
+
     public boolean inTaggedRegion(Location location, Set<String> tags) {
         if (location != null && worldGuardPlugin != null && customFlags != null)
         {