+++ Magic/src/main/java/com/elmakers/mine/bukkit/protection/WorldGuardFlagsManager.java
@@ -33,6 +33,7 @@ public class WorldGuardFlagsManager implements WorldGuardFlags {
     public static StringFlag DESTRUCTIBLE = new StringFlag("destructible", RegionGroup.ALL);
     public static StringFlag REFLECTIVE = new StringFlag("reflective", RegionGroup.ALL);
     public static StringFlag PORTAL_SPELL = new StringFlag("portal-spell", RegionGroup.ALL);
+    public static StringFlag PORTAL_WARP = new StringFlag("portal-warp", RegionGroup.ALL);
 
     // These are outdated but we'll probably keep them forever anyway
     public static SetFlag<String> SPAWN_TAGS = new SetFlag<>("spawn-tags", RegionGroup.ALL, new StringFlag(null));
@@ -67,13 +68,20 @@ public class WorldGuardFlagsManager implements WorldGuardFlags {
         registry.register(TAGS);
         registry.register(SPAWN_TAGS);
         registry.register(PORTAL_SPELL);
-        callingPlugin.getLogger().info("Registered custom WorldGuard flags: allowed-spells, always-allowed-spells, blocked-spells, always-allowed-spell-categories, allowed-spell-categories, blocked-spell-categories, allowed-wands, always-allowed-wands, blocked-wands, spell-overrides, destructible, reflective, tags, portal-spell");
+        registry.register(PORTAL_WARP);
+        callingPlugin.getLogger().info("Registered custom WorldGuard flags: allowed-spells, always-allowed-spells, blocked-spells, always-allowed-spell-categories, allowed-spell-categories, blocked-spell-categories, allowed-wands, always-allowed-wands, blocked-wands, spell-overrides, destructible, reflective, tags, portal-spell, portal-warp");
     }
 
     @Nullable
     @Override
     public String getPortalSpell(RegionAssociable source, ApplicableRegionSet checkSet) {
         return checkSet.queryValue(source, PORTAL_SPELL);
+    }
+
+    @Nullable
+    @Override
+    public String getPortalWarp(RegionAssociable source, ApplicableRegionSet checkSet) {
+        return checkSet.queryValue(source, PORTAL_WARP);
     }
 
     @Nullable