+++ Magic/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Additions
 
- - Add `portal-spell` WorldGuard flag, for making custom portals
+ - Add `portal-spell` and `portal-warp` WorldGuard flags, for making custom portals
 
 # 9.4.1